requirements.txt
arrow==1.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
geojson==2.5.0
idna==3.6
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
paho-mqtt==1.6.1
Pillow==10.1.0
pyowm==3.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
types-python-dateutil==2.8.19.14
urllib3==2.1.0
zipp==3.17.0
